requirements.txt
bcrypt==3.1.7
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Jinja2==2.10.3 
psycopg2==2.8.4 #postgres 접근할때 사용
PyJWT-1.7.1